feat(restaurant-frontend admin-food): get all foods, update food
