+++ apps/L1G/restaurant/restaurant-frontend/specs/(admin)/components/food/FormRadioInput.spec.tsx
@@ -17,7 +17,7 @@ describe('RadioInput', () => {
   it('should render', () => {
     const { getByTestId } = render(<TestWrapper />);
 
-    expect(getByTestId('create-food-activestatus-input')).toBeInTheDocument();
-    expect(getByTestId('create-food-inactivestatus-input')).toBeInTheDocument();
+    expect(getByTestId('food-status-active')).toBeInTheDocument();
+    expect(getByTestId('food-status-inactive')).toBeInTheDocument();
   });
 });
