+++ apps/L1G/restaurant/restaurant-frontend/src/components/admin/index.ts
@@ -1,8 +1,8 @@
 export * from './DeleteDialog';
 export * from './food/FoodCreateDialog';
-export * from './food/FormRadioInput';
 export * from './food/FormSelectInput';
 export * from './food/FormTextInput';
 export * from './food/AllFoodsCard';
 export * from './food/FoodUpdateDialog';
 export * from './food/FoodUpdateForm';
+export * from './food/FormRadioInput';
